last changes

+++ ProyectoPlanetas/src/main/java/com/planetas/Services/PlanetaService.java
@@ -105,16 +105,11 @@ public class PlanetaService implements ObjectServices <PlanetaDTO> {
 	}
 	
 	//----------Delete----------
-	@Transactional
-	public boolean delete(int id) throws Exception{
-		try {
-			if(repository.existsById(id)) {
-				repository.deleteById(id);
-				return true;
-			} else {
-				throw new Exception();
-			}	
-		} catch (Exception e) { throw new Exception();}
-	}
+	  @Transactional
+	  public boolean delete(int id) throws Exception{ try {
+	  if(repository.existsById(id)) { repository.deleteById(id); return true; }
+	  else { throw new Exception(); } } catch (Exception e) { throw new
+	  Exception();} }
+	 
 	
 }
